minor fix: moved components to respective folder
