+++ frontend/pages/community/crm/companies/index.tsx
@@ -4,7 +4,7 @@ import { NextPage } from "next";
 import ContentLayout from "~community/common/components/templates/ContentLayout/ContentLayout";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { IconName } from "~community/common/types/IconTypes";
-import CompanyPopupController from "~community/crm/components/organisms/CompanyPopupController";
+import CompanyPopupController from "~community/crm/components/organisms/CompanyPopupController/CompanyPopupController";
 import { useCrmStore } from "~community/crm/store/crmStore";
 import { CrmModalTypes } from "~community/crm/types/ModalTypes";
